--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,5 @@
+module github.com/dgryski/go-maglev
+
+go 1.19
+
+require github.com/dchest/siphash v1.2.3