Hi. I made some optimizations to the code. The tests are also green. The results of the benchmarks are shown below:

```
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
            │   old.txt   │               new.txt               │
            │   sec/op    │   sec/op     vs base                │
Generate-12   4.488µ ± 1%   4.036µ ± 1%  -10.07% (p=0.000 n=10)
New-12        3.583m ± 1%   1.899m ± 0%  -46.99% (p=0.000 n=10)
geomean       126.8µ        87.55µ       -30.96%

            │   old.txt    │                new.txt                │
            │     B/op     │     B/op      vs base                 │
Generate-12   2.000Ki ± 0%   2.000Ki ± 0%       ~ (p=1.000 n=10) ¹
New-12        523.1Ki ± 0%   523.1Ki ± 0%       ~ (p=0.211 n=10)
geomean       32.35Ki        32.35Ki       +0.00%
¹ all samples are equal

            │  old.txt   │               new.txt               │
            │ allocs/op  │ allocs/op   vs base                 │
Generate-12   2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
New-12        5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean       3.162        3.162       +0.00%
¹ all samples are equal
```